Drop the installed benchmark CLI in favor of the dev-tree script
The torchrl.benchmark package and torchrl-benchmark-rnn console script are
removed: the benchmark stays a developer tool under benchmarks/, carrying
the CLI fixes (module dtype cast, device pinning for timings, --backends
validation, block_gru recompute rejection). The DreamerV3 docs now compare
the reference, scan, and triton backends directly: portability, hardware,
precision and promotion rules, double-backward behavior, and autotune
warmup, with the class docstring pointing at the comparison.
