+++ pyproject.toml
@@ -202,7 +202,6 @@ dev = [
 [project.scripts]
 rlrender = "torchrl.render.cli:main"
 torchrl-render = "torchrl.render.cli:main"
-torchrl-benchmark-rnn = "torchrl.benchmark.cli:main"
 
 [project.urls]
 Homepage = "https://github.com/pytorch/rl"
