--- torchrl/benchmark/__init__.py
@@ -1,6 +1,0 @@
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-#
-# This source code is licensed under the MIT license found in the
-# LICENSE file in the root directory of this source tree.
-"""Hardware benchmarks distributed with TorchRL."""
-from __future__ import annotations
